Factor out BaseTest into common
