+++ test/unit/data/model/mapping/test_model_mapping.py
@@ -68,6 +68,7 @@ from sqlalchemy import func, select
 
 from galaxy import model
 from .common import (
+    AbstractBaseTest,
     collection_consists_of_objects,
     dbcleanup,
     dbcleanup_wrapper,
@@ -80,17 +81,10 @@ from .common import (
 )
 
 
-class BaseTest:
-    @pytest.fixture
-    def cls_(self):
-        """
-        Return class under test.
-        Assumptions: if the class under test is Foo, then the class grouping
-        the tests should be a subclass of BaseTest, named TestFoo.
-        """
-        prefix = len("Test")
-        class_name = self.__class__.__name__[prefix:]
-        return getattr(model, class_name)
+class BaseTest(AbstractBaseTest):
+
+    def get_model(self):
+        return model
 
 
 class TestAPIKeys(BaseTest):
